+++ src/pages/RecrutingPartyPage.jsx
@@ -1,8 +1,7 @@
 import { useState, useCallback } from 'react';
 import { useNavigate, useParams } from 'react-router';
 import { Button, IconButton, Modal } from '@mui/material';
-import { Box } from '@mui/system';
-import { styled } from '@mui/system';
+import ModalBox from 'components/Modal/ModalBox';
 import CloseIcon from '@mui/icons-material/Close';
 
 import PageHeader from 'components/PageHeader';
@@ -110,16 +109,3 @@ const RecrutingPartyPage = () => {
 };
 
 export default RecrutingPartyPage;
-
-const ModalBox = styled(Box)`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  padding: 24px;
-  width: 86%;
-  border-radius: 24px;
-  background-color: #fff;
-  box-shadow: 1px 2px 10px rgba(0, 0, 0, 0.25);
-  text-align: center;
-`;
